Do not set shadowDatabaseUrl to be the same as directUrl, that breaks things
From https://www.prisma.io/docs/orm/prisma-migrate/understanding-prisma-migrate/shadow-database

"Important: Do not use the exact same values for url and shadowDatabaseUrl as that might delete all your database in your database."

This is from the starter code. Seems to work fine if I just remove it and the docs say not specifying it is normal.