Throw when query is 'c:' and similar
The appended space on the query is useful to stop us running off the end of the
string but it makes us see that kind of query as valid when it is not. It might
be nicer to get rid of the space appending but this is simpler for now.